Some documentation, cleanup
